MULE-11295: Mule API Clean up
_ Cleaning Mule Client API
