+++ core/src/main/java/org/mule/runtime/core/api/client/AbstractPriorizableConnectorMessageProcessorProvider.java
@@ -4,9 +4,8 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.core.client;
+package org.mule.runtime.core.api.client;
 
-import org.mule.runtime.core.api.client.AbstractConnectorMessageProcessorProvider;
 import org.mule.runtime.core.api.connector.ConnectorOperationProvider;
 
 /**
