+++ core/src/main/resources/META-INF/mule-module.properties
@@ -51,7 +51,6 @@ artifact.export.classPackages=org.mule.runtime.core,\
                               org.mule.runtime.core.api.transformer.wire,\
                               org.mule.runtime.core.api.transport,\
                               org.mule.runtime.core.api.util,\
-                              org.mule.runtime.core.client,\
                               org.mule.runtime.core.component,\
                               org.mule.runtime.core.component.simple,\
                               org.mule.runtime.core.config,\
